+++ test/transform/resource/after-delombok/FieldDefaults.java
@@ -1,4 +1,5 @@
 class FieldDefaults1 {
+	static int STATIC = 3;
 	final int x;
 	int y;
 	FieldDefaults1(int x) {
@@ -6,6 +7,7 @@ class FieldDefaults1 {
 	}
 }
 class FieldDefaults2 {
+	static int STATIC = 3;
 	int x;
 	private int y;
 }