+++ test/transform/resource/before/FieldDefaults.java
@@ -1,5 +1,6 @@
 @lombok.experimental.FieldDefaults(makeFinal = true)
 class FieldDefaults1 {
+	static int STATIC = 3;
 	int x;
 	@lombok.experimental.NonFinal int y;
 	
@@ -10,6 +11,7 @@ class FieldDefaults1 {
 
 @lombok.experimental.FieldDefaults(level = lombok.AccessLevel.PRIVATE)
 class FieldDefaults2 {
+	static int STATIC = 3;
 	@lombok.experimental.PackagePrivate int x;
 	int y;
 }